adding stream facility test
